refactor for components
