+++ src/components/cards.py
@@ -1,0 +1,27 @@
+import numpy as np
+from typing import Tuple
+
+def pct_on_time_calc(delay_times: np.ndarray) -> Tuple[int, float]:
+
+    # consider delay less than this number to be on time
+    tol = 5
+    if delay_times.size > 0:
+        return (delay_times <= tol).sum() / delay_times.size * 100
+    return 0.0
+
+
+def avg_flight_time(flight_times: np.ndarray) -> Tuple[int, int]:
+    if flight_times.size > 0:
+        m = np.nanmean(flight_times)
+        if isinstance(m, float):
+            hrs = int(m // 60)
+            mins = int(m % 60)
+            return hrs, mins
+    return 0, 0
+
+
+def avg_delay(delay_times: np.ndarray) -> float:
+    if delay_times.size > 0:
+        m = np.nanmean(delay_times)
+        return m
+    return 0
